pyproject.toml
[project]
name = "moondev-clawdbot"
version = "0.1.0"
description = "Local clone of the MoonDev 'clawdbot' social arbitrage pipeline (ingest -> normalize -> score -> alert -> dashboard)."
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "You"}]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "python-dotenv>=1.0.1",
  "pydantic>=2.7.4",
  "requests>=2.32.3",
  "feedparser>=6.0.11",
  "beautifulsoup4>=4.12.3",
  "tqdm>=4.66.4",
  "streamlit>=1.37.0",
  "playwright>=1.42.0",
]

[project.scripts]
moondev-clawdbot = "moondev_clawdbot.cli:app"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/moondev_clawdbot"]
